authorization added for requests

+++ app/Http/Controllers/UsersController.php
@@ -7,13 +7,18 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\User;
 use App\Http\Resources\UserResource;
+use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\StoreRequest;
 use Illuminate\Support\Facades\Validator;
 
 
+
+
 class UsersController extends Controller
 {
+
+
     public function index()
     {
         $user = User::paginate(5);
@@ -28,8 +33,12 @@ class UsersController extends Controller
         $user = User::findOrfail($id);
 
 
+
         // return user as a resource
-        return  new UserResource($user);
+        if(Auth::id() == $id)
+            return  new UserResource($user);
+        else
+            return response(['message' => 'you can only view your own info'] , 401);
     }
 
 
@@ -43,24 +52,46 @@ class UsersController extends Controller
 
     public function update(Request $request , $id)
     {
-
         $user = User::findorfail($id);
+        if(Auth::id() == $id)
+        {
+            $validator = Validator::make($request->all() ,[
+                    'name'=>'required',
+                    'email'=>'required|email|unique:users',
 
-        $user->name = $request->name;
-        $user->email = $request->email;
+                ]);
+            if($validator -> fails())
+            {
+                return response(['errors' , $validator->errors()->all()]  , 400);
+            }
+            $user->name = $request->name;
+            $user->email = $request->email;
 
-        if($user->save()){
-            return new UserResource($user);
+            $user->save();
+
         }
+        else
+        {
+            return response(['message' => 'you can only modify your own info'] , 401);
+        }
+
     }
 
     public function destroy($id)
     {
         // get user
         $user = User::findOrfail($id);
+        if(Auth::id() == $id)
+        {
+            if($user->delete()){
+                return  new UserResource($user);
+            }
+        }
+        else
+        {
+          return response(['message' => 'you can only delete your own info'] , 401);
+        }
 
-       if($user->delete()){
-        return  new UserResource($user);
-       }
+
     }
 }
